Solve dependency issue
