+++ src/main/scala/org/calinburloiu/music/microtuner/tuner/TuningSwitchProcessor.scala
@@ -16,62 +16,8 @@
 
 package org.calinburloiu.music.microtuner.tuner
 
-import com.typesafe.scalalogging.StrictLogging
-import org.calinburloiu.music.microtuner.config.CcTriggers
 import org.calinburloiu.music.scmidi.MidiProcessor
-
-import javax.sound.midi.{MidiMessage, ShortMessage}
-import scala.collection.mutable
 
 trait TuningSwitchProcessor extends MidiProcessor {}
 
-class CcTuningSwitchProcessor(tuningSwitcher: TuningSwitcher,
-                              ccTriggers: CcTriggers) extends TuningSwitchProcessor with StrictLogging {
 
-  private val ccPrev = ccTriggers.prevTuningCc
-  private val ccNext = ccTriggers.nextTuningCc
-  private val ccTriggerThreshold = ccTriggers.ccThreshold
-  private val isFilteringCcTriggersThru: Boolean = ccTriggers.isFilteringThru
-  private val ccDepressed: mutable.Map[Int, Boolean] = mutable.Map(ccPrev -> false, ccNext -> false)
-
-  override def send(message: MidiMessage, timeStamp: Long): Unit = message match {
-    case shortMessage: ShortMessage =>
-      val command = shortMessage.getCommand
-      val cc = shortMessage.getData1
-      val ccValue = shortMessage.getData2
-
-      // Capture Control Change messages used for switching and forward any other message
-      if (command == ShortMessage.CONTROL_CHANGE && ccDepressed.contains(cc)) {
-        if (!ccDepressed(cc) && ccValue > ccTriggerThreshold) {
-          ccDepressed(cc) = true
-          if (cc == ccPrev)
-            tuningSwitcher.prev()
-          else
-            tuningSwitcher.next()
-        } else if (ccDepressed(cc) && ccValue <= ccTriggerThreshold) {
-          ccDepressed(cc) = false
-        }
-
-        // Forward if configured so
-        if (!isFilteringCcTriggersThru) {
-          receiver.send(message, timeStamp)
-        }
-      } else {
-        // Forward
-        receiver.send(message, timeStamp)
-      }
-  }
-
-  override def close(): Unit = {
-    super.close()
-    logger.info(s"Closing ${this.getClass.getCanonicalName}...")
-  }
-
-  override protected def onConnect(): Unit = {
-    logger.info(s"Connected the CC tuning switch MIDI processor.")
-  }
-
-  override protected def onDisconnect(): Unit = {
-    logger.info(s"Disconnected the CC tuning switch MIDI processor.")
-  }
-}
